+++ src/Config/Config.php
@@ -22,6 +22,7 @@ class Config
     const LITHUANIA_ISO_CODE = 'LT';
     const LATVIA_ISO_CODE = 'LV';
     const ESTONIA_ISO_CODE = 'EE';
+    const PORTUGAL_ISO_CODE = 'PT';
 
     const AVAILABLE_PUDO_COD_IDS = [
         'EE90',
@@ -236,7 +237,7 @@ class Config
 
     const PRODUCT_NAME_PUDO = [
         'LT' => 'Pristatymas į paštomatą',
-        'EE' => 'DPD pakiautomaat lunamaksega',
+        'EE' => 'DPD pakiautomaat',
         'LV' => 'Piegādes uz Paku Skapjiem un Paku Bodēm',
         'EN' => 'DPD Pickup points',
     ];
@@ -445,7 +446,7 @@ class Config
                 switch ($productReference) {
                     case self::PRODUCT_TYPE_PUDO:
                     case self::PRODUCT_TYPE_PUDO_COD:
-                        return 20;
+                        return 31.5;
                     case self::PRODUCT_TYPE_SATURDAY_DELIVERY:
                     case self::PRODUCT_TYPE_SAME_DAY_DELIVERY:
                     case self::PRODUCT_TYPE_SATURDAY_DELIVERY_COD:
@@ -458,7 +459,7 @@ class Config
                 switch ($productReference) {
                     case self::PRODUCT_TYPE_PUDO:
                     case self::PRODUCT_TYPE_PUDO_COD:
-                        return 20;
+                        return 31.5;
                     case self::PRODUCT_TYPE_SAME_DAY_DELIVERY:
                         return 31.5;
                     case self::PRODUCT_TYPE_SATURDAY_DELIVERY_COD:
@@ -466,8 +467,30 @@ class Config
                     default:
                         return false;
                 }
+            case self::PORTUGAL_ISO_CODE:
+                switch ($productReference) {
+                    case self::PRODUCT_TYPE_PUDO:
+                    case self::PRODUCT_TYPE_PUDO_COD:
+                        return 10.0;
+                    case self::PRODUCT_TYPE_SATURDAY_DELIVERY:
+                    case self::PRODUCT_TYPE_SAME_DAY_DELIVERY:
+                    case self::PRODUCT_TYPE_SATURDAY_DELIVERY_COD:
+                        return 31.5;
+                    default:
+                        return false;
+                }
             default:
-                return false;
+                switch ($productReference) {
+                    case self::PRODUCT_TYPE_PUDO:
+                    case self::PRODUCT_TYPE_PUDO_COD:
+                        return 20.0;
+                    case self::PRODUCT_TYPE_SATURDAY_DELIVERY:
+                    case self::PRODUCT_TYPE_SAME_DAY_DELIVERY:
+                    case self::PRODUCT_TYPE_SATURDAY_DELIVERY_COD:
+                        return 31.5;
+                    default:
+                        return false;
+                }
         }
     }